pyproject.toml
[tool.poetry]
name = "images"
version = "0.1.0"
description = ""
authors = ["rafiwts <rafiwts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
pre-commit = "^3.4.0"
psycopg2 = "^2.9.7"
python-dotenv = "^1.0.0"
pillow = "^10.0.1"

[tool.ruff]
ignore = ["E501"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
